Case Study 01 - Divvy Bikes in Chicago
Hi there! This is a case study suggested in the last course of the Google Data Analytics course. In this case study, In this case study, yoiu have to annalyze the data that is offered to you to help Divvy Bikes, a share-bike program in Chicago, to turn their casual users into annual users.
After annalyzing the data, I saw that most of users were casual and that most of them prefered to use the electric bike type very often. So, a good marketing campaign to help Divvy to get more annual users, should focus in benefits for annual users, like a percentage discount, mainly in the electric bikes, which is the type of bike that most of users, both annual and casual, prefer. 

A good Marketing Campaign should show the benefits of using an electric bike instead of a car or the bus and the benefits of being a annual user of Divvy Bikes if you need to use an electric bike very often. For example, an electric bike might be to expensive and there are another issues like finding a place to park or someone else steal your bike. So, an interesting idea is to use a bike from a bike-share program, so you don't have to worry about these issues. Other benefits are that electric bikes are cleaner than cars and buses, because most of them produce toxic gases, and that it's also less expensive.

I attached some graphs I created using R Studio and the tables from April 2020 to March 2021 and the code in R. Rate, comment possible insights you may get from this data. The tables are available in the link below to download:
https://divvy-tripdata.s3.amazonaws.com/index.html